BLOG-1 타입 정리 및 기타 수정
