+++ src/app/(__nextjs__)/posts/[id]/page.tsx
@@ -1,5 +1,5 @@
 import { PostPage } from "@/app/_pages/post/ui";
-import { getPost } from "@/app/widgets/post/api/posts";
+import { getPost } from "@/app/_widgets/post/api/posts";
 
 export default async function Page({ params: { id } }: { params: { id: string } }) {
   // 여기서 title, text를 분리..?
